project files restructuration
