+++ src/cell.rs
@@ -1,0 +1,6 @@
+#[derive(Debug, Clone, Copy, PartialEq)]
+
+pub enum Cell{
+    Dead,
+    Alive
+}
